reformat: move userDetails endpoint to UserController with path "/current"
